// src/generated/model/programRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProgramRequestStatus } from './programRequestStatus.ts';

export interface ProgramRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  description?: string;
  status: ProgramRequestStatus;
  displayOrder?: number;
}