// src/generated/model/updateProgramRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateProgramRequestStatus } from './updateProgramRequestStatus.ts';

export interface UpdateProgramRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  description?: string;
  status: UpdateProgramRequestStatus;
  displayOrder?: number;
}